refactor(openehr-adl): regroup validate/ by topic — orchestration-only mod, catalogue, topic modules
Replace the pass-number files (phase1/phase2/phase3/phase_flat/fillers) with
topic modules: catalogue, identification, structure, terminology, bindings,
annotations, source_level, specialisation, slots, flat (rm + conformance
unchanged). validate/mod.rs keeps only ValidationIssue, the six public entry
points and the three phase drivers, whose call order and error gating are the
master08 phase schedule and are byte-for-byte unchanged.

The push method triplicated across the phase-1, specialisation and RM walkers
(and open-coded in the flat-form walk) is now one push_issue helper in mod.rs.
Severity/ValidationCode move to validate::catalogue, validate_fillers to
validate::slots, and the TerminologyResolver seam to validate::bindings; the
consumers are rewritten in-change (no re-exports). Behaviour-preserving: all
264 openehr-adl tests keep their names and pass.
