+++ app/ehrbase/src/service/definition/adl14.rs
@@ -25,8 +25,9 @@ use std::str::FromStr;
 use openehr_adl::adl14::convert::{ConvertConfig, parse_and_convert};
 use openehr_adl::adl14::log::ConversionLog;
 use openehr_adl::error::SyntaxError;
+use openehr_adl::validate::catalogue::Severity;
 use openehr_adl::validate::rm::ProductionRmModel;
-use openehr_adl::validate::{Severity, ValidationIssue, validate_source_adl14};
+use openehr_adl::validate::{ValidationIssue, validate_source_adl14};
 use openehr_base::prelude::ArchetypeId;
 use openehr_its::rest::runtime::ValidationError;
 use uuid::Uuid;
