+++ crates/openehr-adl/src/validate/mod.rs
@@ -1,20 +1,39 @@
-//! The AOM2 validation catalogue.
+//! The AOM2 validation engine: orchestration over a catalogue of topic modules.
 //!
 //! A validator walks an assembled `openehr_am::am24::aom2` [`Archetype`]
-//! and produces typed [`ValidationIssue`]s, each carrying a [`ValidationCode`]
-//! (one variant per AOM2 validity code), a [`Severity`], a message, and — where
-//! derivable — the archetype path the issue is anchored at.
+//! and produces typed [`ValidationIssue`]s, each carrying a
+//! [`ValidationCode`] (one variant per AOM2 validity
+//! code), a [`Severity`], a message, and — where derivable
+//! — the archetype path the issue is anchored at.
 //!
-//! Phase 1 (basic integrity, standalone) is implemented in `phase1`
-//! ([`validate_phase1`] / [`validate_source_phase1`]); the phase-2
-//! reference-model checks are in [`rm`] ([`rm::validate_phase2_rm`]). The
-//! phase-1 catalogue and its orchestration are defined in
+//! This module holds ONLY the orchestration: [`ValidationIssue`], the six public
+//! entry points, and the three phase drivers (`run_phase1`,
+//! `run_phase2_spec`, `run_phase3`) that call the topic modules in the order
+//! the spec's phase schedule prescribes. Every rule lives in the topic module
+//! that owns its subject matter:
+//!
+//! | module | topic |
+//! |---|---|
+//! | [`catalogue`] | the code vocabulary: `Severity` + `ValidationCode` |
+//! | `identification` | archetype id, root typename/concept, versions, languages, the terminology-structure gate |
+//! | `structure` | the phase-1 definition-tree walk (node identity, cardinality, slots, assumed values) |
+//! | `terminology` | term definitions, value sets and the codes the definition references |
+//! | [`bindings`] | binding keys (VTTBK/VTCBK) + the external terminology-service seam (VETDF) |
+//! | `annotations` | the `annotations` and `rm_overlay` sections |
+//! | `source_level` | the checks that read the raw parsed source (VOKU, VRRLP) |
+//! | `specialisation` | a differential child against its flat parent |
+//! | [`slots`] | slot redefinition/filling + the template-filler pass |
+//! | [`rm`] | the reference-model seam and its checks |
+//! | [`conformance`] | the `master04.5` conformance functions the above build on |
+//! | `flat` | everything decidable only on the flattened form |
+//!
+//! The phase-1 catalogue and its orchestration are defined in
 //! `docs/specs/openehr/AM/docs/AOM2/master08-validation.adoc` §Phase 1 - Basic
 //! Integrity, with the full rule texts in `master03-archetype_package.adoc`,
 //! `master04.5-constraint_model-class_definitions.adoc`,
 //! `master06-rm_overlay.adoc`, and `master07-terminology_package.adoc` (cited
-//! per check in `phase1`); the phase-2 RM checks are `master08` §Phase 2 →
-//! Validate Against Reference Model (cited per check in [`rm`]).
+//! per check in the topic modules); the phase-2 RM checks are `master08` §Phase
+//! 2 → Validate Against Reference Model (cited per check in [`rm`]).
 //!
 //! Phase orchestration follows `master08` "multi-pass … more basic kinds of
 //! errors being checked first": phase 2 runs only after phase 1 passes, which
@@ -24,454 +43,38 @@
 //! to the standalone half they can compute (or are skipped).
 //!
 //! The specialised-archetype-vs-flat-parent checks (VSON*/VSANC*/VSSM/VDSS*/
-//! VARX*/…) live in `phase2` and run against the flat parent resolved via
-//! [`resolve_flat_parent`]; the `master04.5` conformance functions they build on
-//! are in [`conformance`]. Per `ADL2/master09.02` §Differential and Flat Forms
-//! ("For a top-level archetype, the flat-form is the same as its differential
-//! form") a level-0 parent is used as-is; a parent that is itself specialised
-//! needs the full flattener before its DEEP flat form is available — which
-//! [`crate::flatten::flat_form`] now supplies, so a specialised parent
-//! ([`FlatParent::NeedsFlattener`]) is flattened before the deep phase-2 checks
-//! run. Phase 3 (`phase3`) runs on the flattened form (VUNP, VACMCO), per
-//! `master08` §Phase 3 - Validation of Flat Form.
+//! VARX*/…) live in `specialisation` (with the slot arm in [`slots`]) and run
+//! against the flat parent resolved via [`resolve_flat_parent`]; the
+//! `master04.5` conformance functions they build on are in [`conformance`]. Per
+//! `ADL2/master09.02` §Differential and Flat Forms ("For a top-level archetype,
+//! the flat-form is the same as its differential form") a level-0 parent is used
+//! as-is; a parent that is itself specialised needs the full flattener before
+//! its DEEP flat form is available — which [`crate::flatten::flat_form`] now
+//! supplies, so a specialised parent ([`FlatParent::NeedsFlattener`]) is
+//! flattened before the deep phase-2 checks run. Phase 3 (`flat`) runs on the
+//! flattened form (VUNP, VACMCO), per `master08` §Phase 3 - Validation of Flat
+//! Form.
 
+mod annotations;
+pub mod bindings;
+pub mod catalogue;
 pub mod conformance;
-pub mod fillers;
-mod phase1;
-mod phase2;
-mod phase3;
-mod phase_flat;
+mod flat;
+mod identification;
 pub mod rm;
-pub mod terminology;
+pub mod slots;
+mod source_level;
+mod specialisation;
+mod structure;
+mod terminology;
 
 use openehr_am::am24::aom2::archetype::archetype::Archetype;
 
-use crate::artefact::{ArchetypeRepository, FlatParent, resolve_flat_parent, view};
+use crate::artefact::{ArchetypeRepository, ArchetypeView, FlatParent, resolve_flat_parent, view};
 use crate::error::SyntaxError;
-use crate::source::{parse_source, parse_source_adl14};
-
-/// The severity of a [`ValidationIssue`].
-///
-/// The `W`-prefixed codes (WACMCL, WOUC) are warnings; every other code is an
-/// error. `master08` assigns no explicit severity column, so this follows the
-/// `V`/`W` naming convention (the `W` prefix = advisory "should"; see
-/// `master04.5` WACMCL "should be" vs VACMCU "must").
-///
-/// NOTE: no openEHR spec states the `W`→Warning convention normatively; it is
-/// inferred from the code naming (`master08-validation` is silent on severity).
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-pub enum Severity {
-    /// A validity error — the archetype is invalid.
-    Error,
-    /// A validity warning — advisory, does not invalidate the archetype.
-    Warning,
-}
-
-/// An AOM2 validation code (one typed variant per validity rule).
-///
-/// Each variant's doc comment names the spec file + section that defines it.
-/// The catalogue is the phase-1 set of `docs/specs/openehr/AM/docs/AOM2/`
-/// plus the two corpus-adjudicated additions (VRDLA, WOUC — archie parity, no
-/// full vendored text, NOTE-flagged). Deferred variants (their check needs the
-/// RM model, the flat parent, or an external terminology service) are present
-/// as the vocabulary but not raised in phase 1 — see `phase1`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
-#[non_exhaustive]
-pub enum ValidationCode {
-    /// VARDT — archetype definition typename validity (`master03` §Validity Rules).
-    Vardt,
-    /// VARCN — archetype concept validity (`master03` §Validity Rules).
-    Varcn,
-    /// STCNT — missing mandatory part, e.g. terminology (`master08` §Phase 1; no
-    /// full vendored text — NOTE-flagged).
-    Stcnt,
-    /// VACSD — archetype concept specialisation depth (`master03` §Validity Rules).
-    Vacsd,
-    /// VOLT — original language available in terminology (`master08` §Phase 1; no
-    /// full vendored text — NOTE-flagged).
-    Volt,
-    /// VARAV — ADL version validity (`master03` §Validity Rules).
-    Varav,
-    /// VARRV — RM release validity (`master03` §Validity Rules).
-    Varrv,
-    /// VOTM — terminology translations validity (`master03` §Validity Rules).
-    Votm,
-    /// VDIFV — differential path only in specialised archetype (`master04.5`
-    /// §`C_ATTRIBUTE`).
-    Vdifv,
-    /// VDIFP — differential path exists in flat parent (`master04.5` §`C_ATTRIBUTE`;
-    /// checked in `phase2` by resolving the differential path against the flat
-    /// parent, with the RM-path half subsumed by that resolution — see [`rm`]).
-    Vdifp,
-    /// VCORM — object constraint type name exists in the RM (`master04.5`
-    /// §`C_OBJECT`; checked in [`rm`]).
-    Vcorm,
-    /// VCORMT — object type conforms to the owning attribute's RM type
-    /// (`master04.5` §`C_OBJECT`; checked in [`rm`]).
-    Vcormt,
-    /// VCARM — attribute name exists on the enclosing RM type (`master04.5`
-    /// §`C_ATTRIBUTE`; checked in [`rm`]).
-    Vcarm,
-    /// VCAEX — attribute existence conforms to the RM existence (`master04.5`
-    /// §`C_ATTRIBUTE`; checked in [`rm`]).
-    Vcaex,
-    /// VCACA — attribute cardinality conforms to the RM cardinality (`master04.5`
-    /// §`C_ATTRIBUTE`; checked in [`rm`]).
-    Vcaca,
-    /// VCAM — attribute single/multiple arity matches the RM (`master04.5`
-    /// §`C_ATTRIBUTE`; checked in [`rm`]).
-    Vcam,
-    /// VCORMEN — enumeration type constraint kind validity: a primitive
-    /// constraint on an enumeration-typed RM slot must match the enumeration's
-    /// underlying primitive (an integer constraint on a string-based
-    /// enumeration, or vice versa, is invalid). `master08` §Phase 2 lists
-    /// (VCORMENV, VCORMENU, VCORMEN) with no full vendored text; the V/U/EN
-    /// partition below is our reading of that gloss against `master04.2`
-    /// §Constraints on Enumeration Types — NOTE-flagged in [`rm`].
-    Vcormen,
-    /// VCORMENV — enumeration integer-value validity: an integer constraint
-    /// value on an integer-based enumeration slot must be a declared literal
-    /// value (`master08` §Phase 2 + `master04.2` §Constraints on Enumeration
-    /// Types; spec-silent full text — NOTE-flagged in [`rm`]).
-    Vcormenv,
-    /// VCORMENU — enumeration string-value validity: a string constraint value
-    /// on a string-based enumeration slot must be a declared literal value
-    /// (`master08` §Phase 2 + `master04.2` §Constraints on Enumeration Types;
-    /// spec-silent full text — NOTE-flagged in [`rm`]).
-    Vcormenu,
-    /// VATCV — terminology code format validity (`master08` §Phase 1; no full
-    /// vendored text — NOTE-flagged).
-    Vatcv,
-    /// VTSD — specialisation level of codes (`master07` §Validity Rules).
-    Vtsd,
-    /// VTLC — terminology language consistency (`master07` §Validity Rules).
-    Vtlc,
-    /// VTTBK — term binding key valid (`master07` §Validity Rules).
-    Vttbk,
-    /// VTCBK — constraint binding key valid (`master07` §Validity Rules).
-    Vtcbk,
-    /// VETDF — external term validity (`master03` §Validity Rules): a code bound
-    /// to an *external* terminology (SNOMED CT, LOINC, …) must exist in that
-    /// terminology. `openehr-adl` (a network-free spec engine) cannot hold a
-    /// live terminology-service client, so the check is threaded through the
-    /// [`terminology::TerminologyResolver`] seam: the application injects a
-    /// resolver over its terminology service and [`validate`] / [`validate_source`]
-    /// consult it. Entry points that take no resolver do not raise VETDF
-    /// (`master03` "subject to tool accessibility; … no verification was
-    /// possible").
-    Vetdf,
-    /// VTVSID — value-set id defined (`master07` §Validity Rules).
-    Vtvsid,
-    /// VTVSMD — value-set members defined (`master07` §Validity Rules).
-    Vtvsmd,
-    /// VTVSUQ — value-set members unique (`master07` §Validity Rules).
-    Vtvsuq,
-    /// VDSEV — slot 'exclude' constraint validity (`master04.5` §`ARCHETYPE_SLOT`).
-    Vdsev,
-    /// VDSIV — slot 'include' constraint validity (`master04.5` §`ARCHETYPE_SLOT`).
-    Vdsiv,
-    /// VARXRA — `C_ARCHETYPE_ROOT` validity set (`master08` §Phase 1; umbrella for
-    /// VARXNC/VARXAV/VARXR — no full vendored text, NOTE-flagged).
-    Varxra,
-    /// VARXNC — `C_ARCHETYPE_ROOT` node-id conformance (`master08` §Phase 1).
-    Varxnc,
-    /// VARXAV — `C_ARCHETYPE_ROOT` archetype-ref validity (`master08` §Phase 1).
-    Varxav,
-    /// VARXR — external reference resolution (`master08` §Phase 2; checked in
-    /// [`fillers`] by resolving each `use_archetype` reference against the
-    /// supplier repository).
-    Varxr,
-    /// VARXTV — `C_ARCHETYPE_ROOT` type validity (`master08` §Phase 1).
-    Varxtv,
-    /// VATID — all definition codes defined in terminology (`master08` §Phase 1;
-    /// no full vendored text — NOTE-flagged).
-    Vatid,
-    /// VATCD — archetype code specialisation level validity (`master03` §Validity
-    /// Rules).
-    Vatcd,
-    /// VATDF — value code (at-code) validity (`master03` §Validity Rules; a
-    /// non-specialised archetype is checked in `phase1`, the specialised
-    /// flat-form half against the flattened terminology in `phase_flat`).
-    Vatdf,
-    /// VACDF — constraint code (ac-code) validity (`master03` §Validity Rules).
-    Vacdf,
-    /// VATDA — value-set assumed value code validity (`master03` §Validity Rules).
-    Vatda,
-    /// VRANP — annotation path valid (`master03` §Validity Rules; the RM-path half
-    /// is a reference-model check, [`rm`]).
-    Vranp,
-    /// VOKU — object key unique in keyed lists (`master03` §Validity Rules).
-    Voku,
-    /// VARID — archetype identifier validity (`master03` §Validity Rules).
-    Varid,
-    /// VDEOL — original language specified (`master03` §Validity Rules).
-    Vdeol,
-    /// VARD — description specified (`master03` §Validity Rules).
-    Vard,
-    /// VASID — specialisation parent identifier validity (`master03` §Validity
-    /// Rules; fires only when the parent is supplied).
-    Vasid,
-    /// VALC — archetype language conformance (`master03` §Validity Rules; fires
-    /// only when the parent is supplied).
-    Valc,
-    /// VTPL — template/filler language consistency (`master03` §Validity Rules;
-    /// checked in [`fillers`] against the resolved, flattened fillers).
-    Vtpl,
-    /// VRRLP — rule path valid (`master03` §Validity Rules; the RM-extension half
-    /// is a reference-model check, [`rm`]).
-    Vrrlp,
-    /// VCOCD — object constraint definition validity (`master04.5` §`C_OBJECT`).
-    Vcocd,
-    /// VCOID — object node identifier present (`master04.5` §`C_OBJECT`).
-    Vcoid,
-    /// VCOSU — object node identifier unique (`master04.5` §`C_OBJECT`).
-    Vcosu,
-    /// VCATU — sibling attribute uniqueness (`master04.5` §`C_COMPLEX_OBJECT`).
-    Vcatu,
-    /// VDFAI — archetype id validity in slot definition (`master04.5`
-    /// §`ARCHETYPE_SLOT`).
-    Vdfai,
-    /// VDFPT — path validity in definition: any path mentioned in the
-    /// definition section must be valid syntactically and valid with respect
-    /// to the hierarchical structure of the definition section (ADL 1.4 only;
-    /// `ADL1.4/master08-adl.adoc` §Definition Section validity rules — the
-    /// AOM2 mirror is [`ValidationCode::Vunp`] on the flat form).
-    Vdfpt,
-    /// VOBAV — object node assumed value validity (`master04.5`
-    /// §`C_PRIMITIVE_OBJECT`).
-    Vobav,
-    /// VRMVP — RM-visibility path validity (`master06` §Validity).
-    Vrmvp,
-    /// VRMVAV — RM-visibility alias validity (`master06` §Validity).
-    Vrmvav,
-    /// VACSO — single-valued attribute child occurrences validity (`master04.5`
-    /// §`C_ATTRIBUTE`).
-    Vacso,
-    /// VACMCU — cardinality/occurrences upper bound validity (`master04.5`
-    /// §`C_ATTRIBUTE`).
-    Vacmcu,
-    /// VCOC — ADL 1.4 cardinality/occurrences validity: the interval formed by the
-    /// sums of the children's occurrences minima..maxima must be inside the
-    /// container's cardinality interval (`ADL1.4/master05-cadl.adoc` §Occurrences
-    /// L321-324). The ADL 1.4 formalism's own rule, raised only on the 1.4
-    /// dialect; the AOM2 successor is the VACMCU/WACMCL pair.
-    Vcoc,
-    /// VACMCO — cardinality/occurrences orphans: every mandatory child and one
-    /// optional child must fit within the container cardinality (`master04.5`
-    /// §`C_ATTRIBUTE` VACMCO L158-159; a phase-3 flat-form check, `phase3`).
-    Vacmco,
-    /// VSONIF — object node identification validity in flat siblings (`master04.5`
-    /// §`C_OBJECT` VSONIF L356-357; refs the spec-undefined VACMI). Checked in
-    /// `phase2`: a new object node in a specialised container whose flattened
-    /// siblings are identified must itself be identified.
-    Vsonif,
-    /// VRDLA — resource-description language-code consistency (archie parity; no
-    /// openEHR spec governs this — our own design/extension, NOTE-flagged).
-    Vrdla,
-    /// WACMCL — cardinality/occurrences lower bound warning (`master04.5`
-    /// §`C_ATTRIBUTE`; WARNING).
-    Wacmcl,
-    /// WOUC — defined terminology code unused in the definition (archie parity;
-    /// no openEHR spec governs this — our own design/extension; WARNING).
-    Wouc,
-    // ── phase-2 specialisation-vs-flat-parent codes (`master04.5` §Validity
-    //    Rules: `C_ATTRIBUTE` / `C_OBJECT` / `ARCHETYPE_SLOT` / `C_ARCHETYPE_ROOT` /
-    //    `C_COMPLEX_OBJECT_PROXY`; `master08` §Phase 2 → Validate Specialised
-    //    Definition). Raised by `phase2` against the flat parent.
-    /// VSANCE — specialised attribute node existence conformance (`master04.5`
-    /// §`C_ATTRIBUTE`).
-    Vsance,
-    /// VSANCC — specialised attribute node cardinality conformance (`master04.5`
-    /// §`C_ATTRIBUTE`).
-    Vsancc,
-    /// VSAM — specialised attribute multiplicity conformance (`master04.5`
-    /// §`C_ATTRIBUTE`).
-    Vsam,
-    /// VSONIN — new object node identifier validity (`master04.5` §`C_OBJECT`).
-    Vsonin,
-    /// VSSM — specialised sibling order validity (`master04.5` §`C_OBJECT`).
-    Vssm,
-    /// VSONT — specialised object node meta-type conformance (`master04.5`
-    /// §`C_OBJECT`).
-    Vsont,
-    /// VSONCT — specialised object node reference type conformance (`master04.5`
-    /// §`C_OBJECT`).
-    Vsonct,
-    /// VSONCO — specialised object node occurrences redefinition validity
-    /// (`master04.5` §`C_OBJECT` — the collective-occurrences rule).
-    Vsonco,
-    /// VSONPT — prohibited object node AOM type validity (`master04.5` §`C_OBJECT`).
-    Vsonpt,
-    /// VSONPI — prohibited object node node-id validity (`master04.5` §`C_OBJECT`).
-    Vsonpi,
-    /// VSONPO — new object node prohibited occurrences validity (`master04.5`
-    /// §`C_OBJECT`).
-    Vsonpo,
-    /// VSONI — _deprecated_ redefined object node identifier validity (`master04.5`
-    /// §`C_OBJECT`; recognise, do not enforce).
-    Vsoni,
-    /// VSONIR — _deprecated_ redefined object node identifier condition
-    /// (`master04.5` §`C_OBJECT`; recognise, do not enforce).
-    Vsonir,
-    /// VSUNT — `use_node` specialisation parent validity (`master04.5`
-    /// §`C_COMPLEX_OBJECT_PROXY`).
-    Vsunt,
-    /// VUNT — `use_node` reference model type validity (`master04.5`
-    /// §`C_COMPLEX_OBJECT_PROXY`).
-    Vunt,
-    /// VUNP — `use_node` path validity: the proxy target path must resolve to an
-    /// object node in the flat form (`master04.5` §`C_COMPLEX_OBJECT_PROXY`
-    /// VUNP L482-483; a phase-3 flat-form check, `phase3`).
-    Vunp,
-    /// VDSSID — slot redefinition child node id (`master04.5` §`ARCHETYPE_SLOT`).
-    Vdssid,
-    /// VDSSM — specialised slot definition match validity (`master04.5`
-    /// §`ARCHETYPE_SLOT`).
-    Vdssm,
-    /// VDSSP — specialised slot definition parent validity (`master04.5`
-    /// §`ARCHETYPE_SLOT`).
-    Vdssp,
-    /// VDSSC — specialised slot definition closed validity (`master04.5`
-    /// §`ARCHETYPE_SLOT`).
-    Vdssc,
-    /// VARXS — external reference slot conformance (`master04.5`
-    /// §`C_ARCHETYPE_ROOT`).
-    Varxs,
-    /// VARXID — external reference slot filling id validity (`master04.5`
-    /// §`C_ARCHETYPE_ROOT`).
-    Varxid,
-    /// VPOV — invalid leaf object value redefinition (`master08` §Phase 2; no full
-    /// vendored text — implemented from the gloss via `c_value_conforms_to`,
-    /// NOTE-flagged).
-    Vpov,
-    /// VUNK — invalid leaf object value redefinition (`master08` §Phase 2; no full
-    /// vendored text — NOTE-flagged).
-    Vunk,
-    /// VTPNC — tuple non-conformance to the parent node (`master08` §Phase 2; no
-    /// full vendored text — NOTE-flagged).
-    Vtpnc,
-    /// VTPIN — tuple invalidity against the parent node (`master08` §Phase 2; no
-    /// full vendored text — NOTE-flagged).
-    Vtpin,
-}
-
-impl ValidationCode {
-    /// The bare mnemonic (e.g. `"VARDT"`), as used in the spec catalogue and
-    /// the conformance-corpus `regression` tags.
-    #[must_use]
-    pub fn mnemonic(self) -> &'static str {
-        match self {
-            Self::Vardt => "VARDT",
-            Self::Varcn => "VARCN",
-            Self::Stcnt => "STCNT",
-            Self::Vacsd => "VACSD",
-            Self::Volt => "VOLT",
-            Self::Varav => "VARAV",
-            Self::Varrv => "VARRV",
-            Self::Votm => "VOTM",
-            Self::Vdifv => "VDIFV",
-            Self::Vdifp => "VDIFP",
-            Self::Vcorm => "VCORM",
-            Self::Vcormt => "VCORMT",
-            Self::Vcarm => "VCARM",
-            Self::Vcaex => "VCAEX",
-            Self::Vcaca => "VCACA",
-            Self::Vcam => "VCAM",
-            Self::Vcormen => "VCORMEN",
-            Self::Vcormenv => "VCORMENV",
-            Self::Vcormenu => "VCORMENU",
-            Self::Vatcv => "VATCV",
-            Self::Vtsd => "VTSD",
-            Self::Vtlc => "VTLC",
-            Self::Vttbk => "VTTBK",
-            Self::Vtcbk => "VTCBK",
-            Self::Vetdf => "VETDF",
-            Self::Vtvsid => "VTVSID",
-            Self::Vtvsmd => "VTVSMD",
-            Self::Vtvsuq => "VTVSUQ",
-            Self::Vdsev => "VDSEV",
-            Self::Vdsiv => "VDSIV",
-            Self::Varxra => "VARXRA",
-            Self::Varxnc => "VARXNC",
-            Self::Varxav => "VARXAV",
-            Self::Varxr => "VARXR",
-            Self::Varxtv => "VARXTV",
-            Self::Vatid => "VATID",
-            Self::Vatcd => "VATCD",
-            Self::Vatdf => "VATDF",
-            Self::Vacdf => "VACDF",
-            Self::Vatda => "VATDA",
-            Self::Vranp => "VRANP",
-            Self::Voku => "VOKU",
-            Self::Varid => "VARID",
-            Self::Vdeol => "VDEOL",
-            Self::Vard => "VARD",
-            Self::Vasid => "VASID",
-            Self::Valc => "VALC",
-            Self::Vtpl => "VTPL",
-            Self::Vrrlp => "VRRLP",
-            Self::Vcocd => "VCOCD",
-            Self::Vcoid => "VCOID",
-            Self::Vcosu => "VCOSU",
-            Self::Vcatu => "VCATU",
-            Self::Vdfai => "VDFAI",
-            Self::Vdfpt => "VDFPT",
-            Self::Vobav => "VOBAV",
-            Self::Vrmvp => "VRMVP",
-            Self::Vrmvav => "VRMVAV",
-            Self::Vacso => "VACSO",
-            Self::Vacmcu => "VACMCU",
-            Self::Vcoc => "VCOC",
-            Self::Vacmco => "VACMCO",
-            Self::Vsonif => "VSONIF",
-            Self::Vrdla => "VRDLA",
-            Self::Wacmcl => "WACMCL",
-            Self::Wouc => "WOUC",
-            Self::Vsance => "VSANCE",
-            Self::Vsancc => "VSANCC",
-            Self::Vsam => "VSAM",
-            Self::Vsonin => "VSONIN",
-            Self::Vssm => "VSSM",
-            Self::Vsont => "VSONT",
-            Self::Vsonct => "VSONCT",
-            Self::Vsonco => "VSONCO",
-            Self::Vsonpt => "VSONPT",
-            Self::Vsonpi => "VSONPI",
-            Self::Vsonpo => "VSONPO",
-            Self::Vsoni => "VSONI",
-            Self::Vsonir => "VSONIR",
-            Self::Vsunt => "VSUNT",
-            Self::Vunt => "VUNT",
-            Self::Vunp => "VUNP",
-            Self::Vdssid => "VDSSID",
-            Self::Vdssm => "VDSSM",
-            Self::Vdssp => "VDSSP",
-            Self::Vdssc => "VDSSC",
-            Self::Varxs => "VARXS",
-            Self::Varxid => "VARXID",
-            Self::Vpov => "VPOV",
-            Self::Vunk => "VUNK",
-            Self::Vtpnc => "VTPNC",
-            Self::Vtpin => "VTPIN",
-        }
-    }
-
-    /// The severity of this code: [`Severity::Warning`] for the `W`-prefixed
-    /// codes, [`Severity::Error`] otherwise (see [`Severity`]).
-    #[must_use]
-    pub fn severity(self) -> Severity {
-        match self {
-            Self::Wacmcl | Self::Wouc => Severity::Warning,
-            _ => Severity::Error,
-        }
-    }
-}
-
-impl std::fmt::Display for ValidationCode {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.write_str(self.mnemonic())
-    }
-}
+use crate::parse::Dialect;
+use crate::source::{SourceArtefact, parse_source, parse_source_adl14};
+use crate::validate::catalogue::{Severity, ValidationCode};
 
 /// A single validation finding.
 #[derive(Debug, Clone, PartialEq, Eq)]
@@ -508,6 +111,103 @@ impl ValidationIssue {
     }
 }
 
+/// Append a path-anchored issue to `issues` — the one construction site every
+/// walker in this module tree uses, so message + path shaping is identical
+/// across the phase-1, specialisation, reference-model and flat-form walks.
+pub(super) fn push_issue(
+    issues: &mut Vec<ValidationIssue>,
+    code: ValidationCode,
+    msg: impl Into<String>,
+    path: &str,
+) {
+    issues.push(ValidationIssue::new(code, msg).at_path(path.to_owned()));
+}
+
+/// Run the phase-1 catalogue over `v`, appending issues to `issues`.
+///
+/// The order follows `master08-validation.adoc` §Phase 1 - Basic Integrity:
+/// basic identification checks first, then structural, then terminology (the
+/// latter gated behind a clean terminology structure and no basic error —
+/// master08 "basic errors first", and a code cannot be checked against a
+/// missing/inconsistent terminology).
+///
+/// `dialect` selects the validity catalogue: [`Dialect::Adl2`] runs the full
+/// AOM2 phase-1 catalogue; [`Dialect::Adl14`] runs the subset that corresponds
+/// to the ADL 1.4 / AOM 1.4 standalone validity rules (see
+/// [`validate_source_phase1_adl14`] for the correspondence + the suppressed
+/// AOM2-only rules, each spec-cited at its check site).
+///
+/// Not run in phase 1 (the variant is present as the catalogue vocabulary): the
+/// reference-model checks live in [`rm`]; VDIFP + VSONIF against the flat parent
+/// in `specialisation`; the flat-form terminology/structure halves
+/// (VATDF/VTVSMD/VACMCU/VCOSU for a specialised archetype) in `flat`; the
+/// external-reference resolution half of VARXR in [`slots`]; and the pure
+/// reference-model path halves of VRANP/VRRLP/VRMVP (a reference-model path
+/// walk, [`rm`]). VETDF (a code bound to an external terminology must exist
+/// there) needs a live terminology-service resolver the network-free spec engine
+/// cannot hold — it is validated through the [`bindings::TerminologyResolver`]
+/// seam.
+fn run_phase1(
+    v: &ArchetypeView<'_>,
+    repo: Option<&ArchetypeRepository>,
+    source: Option<(&SourceArtefact, &str)>,
+    dialect: Dialect,
+    issues: &mut Vec<ValidationIssue>,
+) {
+    // ── basic identification / meta-data checks (master08 §Basic checks +
+    //    §AUTHORED_ARCHETYPE meta-data checks) ──────────────────────────────
+    let mut basic = Vec::new();
+    identification::check_identification(v, repo, dialect, &mut basic);
+
+    // ── terminology structure (STCNT / VOLT) — gates the code checks ───────
+    let term_status = identification::terminology_structure(v);
+    match term_status {
+        identification::TermStructure::Empty => {
+            // STCNT: any missing mandatory part, e.g. the `terminology` section
+            // (master08 §Basic checks; no full vendored text — NOTE-flagged).
+            if v.kind != crate::source::ArtefactKind::TemplateOverlay {
+                basic.push(ValidationIssue::new(
+                    ValidationCode::Stcnt,
+                    "the terminology section defines no term_definitions",
+                ));
+            }
+        }
+        identification::TermStructure::MissingOriginalLanguage => {
+            // VOLT: original language available in the terminology section
+            // (master08 §AUTHORED_ARCHETYPE meta-data checks; NOTE-flagged).
+            issues.push(ValidationIssue::new(
+                ValidationCode::Volt,
+                format!(
+                    "the original language {:?} has no term_definitions bucket",
+                    identification::original_language(v)
+                ),
+            ));
+        }
+        identification::TermStructure::Ok => {}
+    }
+
+    // ── structural definition walk (always runs; independent rules) ────────
+    structure::check_structure(v, dialect, issues);
+    annotations::check_annotations(v, issues);
+    annotations::check_rm_overlay(v, issues);
+    identification::check_resource_description_languages(v, issues); // VRDLA
+    if let Some((src, text)) = source {
+        source_level::check_object_key_unique(src, issues); // VOKU (source-level)
+        source_level::check_rule_paths(v, src, text, issues); // VRRLP (raw rules text)
+        if dialect == Dialect::Adl14 {
+            identification::check_concept_term_adl14(v, src, issues); // VARCN (terminology half)
+        }
+    }
+
+    let basic_clean = basic.is_empty();
+    issues.append(&mut basic);
+
+    // ── terminology + code checks (gated: basic clean + terminology Ok) ────
+    if basic_clean && term_status == identification::TermStructure::Ok {
+        terminology::check_terminology(v, dialect, issues);
+    }
+}
+
 /// Validate an assembled [`Archetype`] against the AOM2 **phase-1** catalogue.
 ///
 /// `repo`, when supplied, resolves the archetype's parent (and suppliers) so
@@ -521,13 +221,7 @@ pub fn validate_phase1(
     repo: Option<&ArchetypeRepository>,
 ) -> Vec<ValidationIssue> {
     let mut issues = Vec::new();
-    phase1::run(
-        &view(archetype),
-        repo,
-        None,
-        crate::parse::Dialect::Adl2,
-        &mut issues,
-    );
+    run_phase1(&view(archetype), repo, None, Dialect::Adl2, &mut issues);
     issues
 }
 
@@ -545,11 +239,11 @@ pub fn validate_source_phase1(
     let source = parse_source(src)?;
     let archetype = crate::assemble::assemble(&source, src)?;
     let mut issues = Vec::new();
-    phase1::run(
+    run_phase1(
         &view(&archetype),
         repo,
         Some((&source, src)),
-        crate::parse::Dialect::Adl2,
+        Dialect::Adl2,
         &mut issues,
     );
     Ok(issues)
@@ -564,7 +258,8 @@ pub fn validate_source_phase1(
 /// source is validated against the 1.4 formalism's own (smaller) catalogue. The
 /// checks that correspond to an ADL 1.4 / AOM 1.4 rule run unchanged; the
 /// AOM2-only rules that would false-reject a valid 1.4 archetype are suppressed
-/// at their check sites in `phase1` (each spec-cited there):
+/// at their check sites in `identification` / `structure` / `terminology`
+/// (each spec-cited there):
 /// - **VARAV / VARRV** — AOM 1.4 has no `adl_version`/`rm_release` 3-part rule
 ///   (`adl_version` is `1.4`-form metadata; 1.4 carries no `rm_release`).
 /// - **VCOID** — relaxed to the AOM 1.4 `node_id` rule (required only for
@@ -585,8 +280,8 @@ pub fn validate_source_phase1(
 /// Two checks are 1.4-ONLY (the ADL 1.4 formalism's own rules, absent from AOM2):
 /// - **VCOC** — cardinality/occurrences validity over the children's EFFECTIVE
 ///   occurrences (`ADL1.4/master05-cadl.adoc` §Occurrences L321-324; the AOM2
-///   successor is the VACMCU/WACMCL pair). See `phase1` for the adjudication of
-///   which half of the literal formula is enforceable.
+///   successor is the VACMCU/WACMCL pair). See `structure` for the
+///   adjudication of which half of the literal formula is enforceable.
 /// - **VATDF/VACDF over the 1.4 term-constraint spelling** — the qualified/listed
 ///   form `[local:: a, b ; assumed]` of
 ///   `ADL1.4/master09-customising_adl.adoc` §Custom Syntax is decomposed into its
@@ -595,7 +290,7 @@ pub fn validate_source_phase1(
 /// - **VDFPT** — `use_node` target paths must resolve within the definition
 ///   section (`ADL1.4/master08-adl.adoc` §Definition Section; a 1.4 artefact
 ///   is standalone, so its own assembled definition is the resolution target —
-///   `phase3::validate_definition_paths_adl14`).
+///   `flat::validate_definition_paths_adl14`).
 ///
 /// # Errors
 /// Returns the parse [`SyntaxError`]s if `src` does not parse as ADL 1.4;
@@ -604,14 +299,14 @@ pub fn validate_source_phase1_adl14(src: &str) -> Result<Vec<ValidationIssue>, V
     let source = parse_source_adl14(src)?;
     let archetype = crate::assemble::assemble_adl14(&source, src)?;
     let mut issues = Vec::new();
-    phase1::run(
+    run_phase1(
         &view(&archetype),
         None,
         Some((&source, src)),
-        crate::parse::Dialect::Adl14,
+        Dialect::Adl14,
         &mut issues,
     );
-    issues.extend(phase3::validate_definition_paths_adl14(&archetype));
+    issues.extend(flat::validate_definition_paths_adl14(&archetype));
     Ok(issues)
 }
 
@@ -622,11 +317,12 @@ pub fn validate_source_phase1_adl14(src: &str) -> Result<Vec<ValidationIssue>, V
 /// VUNT is a rule of the ADL 1.4 formalism itself — `ADL1.4/master05-cadl.adoc`
 /// §Internal References L512-513 — so a 1.4 artefact that violates it is
 /// invalid 1.4, and a 1.4 upload path that stops at phase 1 can never reach it.
-/// The RM pass runs only when phase 1 raised no [`Severity::Error`] — the
-/// `master08` §Overview phase gate ("more basic kinds of errors being checked
-/// first"). A type `rm` does not know is undecidable rather than wrong
-/// ([`rm::type_conforms`] returns `None`), so an artefact built on a reference
-/// model the supplied [`rm::RmModel`] does not carry simply raises no VUNT.
+/// The RM pass runs only when phase 1 raised no
+/// [`Severity::Error`] — the `master08` §Overview
+/// phase gate ("more basic kinds of errors being checked first"). A type `rm`
+/// does not know is undecidable rather than wrong ([`rm::type_conforms`] returns
+/// `None`), so an artefact built on a reference model the supplied
+/// [`rm::RmModel`] does not carry simply raises no VUNT.
 ///
 /// # Errors
 /// Returns the parse [`SyntaxError`]s if `src` does not parse as ADL 1.4;
@@ -638,14 +334,14 @@ pub fn validate_source_adl14(
     let source = parse_source_adl14(src)?;
     let archetype = crate::assemble::assemble_adl14(&source, src)?;
     let mut issues = Vec::new();
-    phase1::run(
+    run_phase1(
         &view(&archetype),
         None,
         Some((&source, src)),
-        crate::parse::Dialect::Adl14,
+        Dialect::Adl14,
         &mut issues,
     );
-    issues.extend(phase3::validate_definition_paths_adl14(&archetype));
+    issues.extend(flat::validate_definition_paths_adl14(&archetype));
     if issues.iter().all(|i| i.severity != Severity::Error) {
         issues.extend(rm::validate_adl14_rm(&archetype, rm));
     }
@@ -653,8 +349,8 @@ pub fn validate_source_adl14(
 }
 
 /// Validate an assembled [`Archetype`] against phase 1 and — only if phase 1
-/// raised no [`Severity::Error`] — the phase-2 reference-model checks against
-/// `rm`.
+/// raised no [`Severity::Error`] — the phase-2
+/// reference-model checks against `rm`.
 ///
 /// This is the `master08` §Overview phase gate ("more basic kinds of errors
 /// being checked first"): the RM pass runs on a structurally-sound archetype
@@ -662,7 +358,7 @@ pub fn validate_source_adl14(
 /// [`validate_source`] for those.
 ///
 /// `resolver` verifies external term bindings (VETDF); pass
-/// [`terminology::NoTerminologyResolver`] when no terminology service is
+/// [`bindings::NoTerminologyResolver`] when no terminology service is
 /// available (VETDF is then not raised — `master03` §Validity Rules "subject to
 /// tool accessibility").
 #[must_use]
@@ -670,7 +366,7 @@ pub fn validate(
     archetype: &Archetype,
     repo: Option<&ArchetypeRepository>,
     rm: &dyn rm::RmModel,
-    resolver: &dyn terminology::TerminologyResolver,
+    resolver: &dyn bindings::TerminologyResolver,
 ) -> Vec<ValidationIssue> {
     let mut issues = validate_phase1(archetype, repo);
     if issues.iter().all(|i| i.severity != Severity::Error) {
@@ -678,7 +374,7 @@ pub fn validate(
     }
     run_phase2_spec(archetype, repo, rm, &mut issues);
     run_phase3(archetype, repo, &mut issues);
-    terminology::check_external_term_bindings(&view(archetype), resolver, &mut issues);
+    bindings::check_external_term_bindings(&view(archetype), resolver, &mut issues);
     issues
 }
 
@@ -701,7 +397,9 @@ fn run_phase2_spec(
     };
     match resolve_flat_parent(archetype, repo) {
         FlatParent::Available(parent) => {
-            issues.extend(phase2::validate_phase2_spec(archetype, parent, rm, repo));
+            issues.extend(specialisation::validate_phase2_spec(
+                archetype, parent, rm, repo,
+            ));
         }
         FlatParent::NeedsFlattener => {
             // The declared parent is itself specialised: flatten it to its deep
@@ -711,7 +409,7 @@ fn run_phase2_spec(
                 && let Some(parent) = repo.get(parent_id)
                 && let Ok(flat_parent) = crate::flatten::flat_form(parent, repo)
             {
-                issues.extend(phase2::validate_phase2_spec(
+                issues.extend(specialisation::validate_phase2_spec(
                     archetype,
                     &flat_parent,
                     rm,
@@ -743,13 +441,13 @@ fn run_phase3(
         None if view(archetype).parent_archetype_id.is_none() => archetype,
         None => return,
     };
-    issues.extend(phase3::validate_phase3(flat_ref));
+    issues.extend(flat::validate_phase3(flat_ref));
     // The deferred flat-form terminology/structure checks (VATDF / VTVSMD /
     // VACMCU / WACMCL / VCOSU) run only for a *specialised* archetype: a
-    // non-specialised archetype is its own flat form, so `phase1` already ran
-    // their equivalents (never double-firing here).
+    // non-specialised archetype is its own flat form, so the phase-1 topic
+    // modules already ran their equivalents (never double-firing here).
     if view(archetype).is_specialised() {
-        issues.extend(phase_flat::validate_flat_form(flat_ref));
+        issues.extend(flat::validate_flat_form(flat_ref));
     }
 }
 
@@ -758,7 +456,7 @@ fn run_phase3(
 /// checks against `rm` (`master08` §Overview phase gate).
 ///
 /// `resolver` verifies external term bindings (VETDF); pass
-/// [`terminology::NoTerminologyResolver`] when no terminology service is
+/// [`bindings::NoTerminologyResolver`] when no terminology service is
 /// available (VETDF is then not raised — `master03` §Validity Rules "subject to
 /// tool accessibility").
 ///
@@ -769,139 +467,26 @@ pub fn validate_source(
     src: &str,
     repo: Option<&ArchetypeRepository>,
     rm: &dyn rm::RmModel,
-    resolver: &dyn terminology::TerminologyResolver,
+    resolver: &dyn bindings::TerminologyResolver,
 ) -> Result<Vec<ValidationIssue>, Vec<SyntaxError>> {
     let source = parse_source(src)?;
     let archetype = crate::assemble::assemble(&source, src)?;
     let mut issues = Vec::new();
-    phase1::run(
+    run_phase1(
         &view(&archetype),
         repo,
         Some((&source, src)),
-        crate::parse::Dialect::Adl2,
+        Dialect::Adl2,
         &mut issues,
     );
     if issues.iter().all(|i| i.severity != Severity::Error) {
         issues.extend(rm::validate_phase2_rm(&archetype, rm));
     }
     run_phase2_spec(&archetype, repo, rm, &mut issues);
-    terminology::check_external_term_bindings(&view(&archetype), resolver, &mut issues);
+    // TODO: run the phase-3 flat-form checks here too — this source-level entry
+    // omits `run_phase3`, which the assembled-archetype entry `validate` runs,
+    // so VUNP/VACMCO (and the deferred flat-form halves) are not raised on the
+    // parse-and-validate path; tracked as issue #1341.
+    bindings::check_external_term_bindings(&view(&archetype), resolver, &mut issues);
     Ok(issues)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    #[expect(
-        clippy::too_many_lines,
-        reason = "one line per catalogue code — an exhaustive list whose whole point is to name every code, not logic"
-    )]
-    fn every_code_has_a_unique_mnemonic_and_severity() {
-        let all = [
-            ValidationCode::Vardt,
-            ValidationCode::Varcn,
-            ValidationCode::Stcnt,
-            ValidationCode::Vacsd,
-            ValidationCode::Volt,
-            ValidationCode::Varav,
-            ValidationCode::Varrv,
-            ValidationCode::Votm,
-            ValidationCode::Vdifv,
-            ValidationCode::Vdifp,
-            ValidationCode::Vcorm,
-            ValidationCode::Vcormt,
-            ValidationCode::Vcarm,
-            ValidationCode::Vcaex,
-            ValidationCode::Vcaca,
-            ValidationCode::Vcam,
-            ValidationCode::Vcormen,
-            ValidationCode::Vcormenv,
-            ValidationCode::Vcormenu,
-            ValidationCode::Vatcv,
-            ValidationCode::Vtsd,
-            ValidationCode::Vtlc,
-            ValidationCode::Vttbk,
-            ValidationCode::Vtcbk,
-            ValidationCode::Vetdf,
-            ValidationCode::Vtvsid,
-            ValidationCode::Vtvsmd,
-            ValidationCode::Vtvsuq,
-            ValidationCode::Vdsev,
-            ValidationCode::Vdsiv,
-            ValidationCode::Varxra,
-            ValidationCode::Varxnc,
-            ValidationCode::Varxav,
-            ValidationCode::Varxr,
-            ValidationCode::Varxtv,
-            ValidationCode::Vatid,
-            ValidationCode::Vatcd,
-            ValidationCode::Vatdf,
-            ValidationCode::Vacdf,
-            ValidationCode::Vatda,
-            ValidationCode::Vranp,
-            ValidationCode::Voku,
-            ValidationCode::Varid,
-            ValidationCode::Vdeol,
-            ValidationCode::Vard,
-            ValidationCode::Vasid,
-            ValidationCode::Valc,
-            ValidationCode::Vtpl,
-            ValidationCode::Vrrlp,
-            ValidationCode::Vcocd,
-            ValidationCode::Vcoid,
-            ValidationCode::Vcosu,
-            ValidationCode::Vcatu,
-            ValidationCode::Vdfai,
-            ValidationCode::Vobav,
-            ValidationCode::Vrmvp,
-            ValidationCode::Vrmvav,
-            ValidationCode::Vacso,
-            ValidationCode::Vacmcu,
-            ValidationCode::Vcoc,
-            ValidationCode::Vacmco,
-            ValidationCode::Vsonif,
-            ValidationCode::Vrdla,
-            ValidationCode::Wacmcl,
-            ValidationCode::Wouc,
-            ValidationCode::Vsance,
-            ValidationCode::Vsancc,
-            ValidationCode::Vsam,
-            ValidationCode::Vsonin,
-            ValidationCode::Vssm,
-            ValidationCode::Vsont,
-            ValidationCode::Vsonct,
-            ValidationCode::Vsonco,
-            ValidationCode::Vsonpt,
-            ValidationCode::Vsonpi,
-            ValidationCode::Vsonpo,
-            ValidationCode::Vsoni,
-            ValidationCode::Vsonir,
-            ValidationCode::Vsunt,
-            ValidationCode::Vunt,
-            ValidationCode::Vunp,
-            ValidationCode::Vdssid,
-            ValidationCode::Vdssm,
-            ValidationCode::Vdssp,
-            ValidationCode::Vdssc,
-            ValidationCode::Varxs,
-            ValidationCode::Varxid,
-            ValidationCode::Vpov,
-            ValidationCode::Vunk,
-            ValidationCode::Vtpnc,
-            ValidationCode::Vtpin,
-        ];
-        let mut seen = std::collections::HashSet::new();
-        for c in all {
-            assert!(seen.insert(c.mnemonic()), "duplicate mnemonic {c}");
-            let expected = if c.mnemonic().starts_with('W') {
-                Severity::Warning
-            } else {
-                Severity::Error
-            };
-            assert_eq!(c.severity(), expected, "{c} severity");
-        }
-        assert_eq!(seen.len(), 91);
-    }
 }
