+++ crates/openehr-adl/tests/it/adl14_custom_constraints.rs
@@ -36,7 +36,8 @@ use openehr_adl::adl14::convert::{ConvertConfig, parse_and_convert};
 use openehr_adl::adl14::log::ConversionLog;
 use openehr_adl::assemble::{parse_artefact, parse_artefact_adl14};
 use openehr_adl::error::{SyntaxError, SyntaxErrorCode};
-use openehr_adl::validate::{Severity, ValidationCode, validate_source_phase1_adl14};
+use openehr_adl::validate::catalogue::{Severity, ValidationCode};
+use openehr_adl::validate::validate_source_phase1_adl14;
 use openehr_am::am24::aom2::archetype::archetype::Archetype;
 use openehr_am::am24::aom2::archetype::authored_archetype::AuthoredArchetype;
 use openehr_am::am24::aom2::constraint_model::c_attribute_tuple::CAttributeTuple;
